// portfolio/src/lib/types.ts
export type ProjectTypeColor =
  | "primary"
  | "secondary"
  | "warning"
  | "info"
  | "success"
  | "default"
  | "error";

export type ProjectStatusColor =
  | "primary"
  | "secondary"
  | "warning"
  | "info"
  | "success"
  | "default"
  | "error";